chore: rename to release-trigger
