+++ packages/release-trigger/.eslintignore
