+++ packages/release-trigger/.eslintrc.json
