+++ packages/release-trigger/cloudbuild-test.yaml
@@ -16,14 +16,14 @@ steps:
   - name: gcr.io/cloud-builders/docker
     id: "build-docker"
     waitFor: ["-"]
-    dir: packages/kokoro-trigger
+    dir: packages/release-trigger
     args:
       - "build"
       - "-t"
-      - "gcr.io/$PROJECT_ID/kokoro-trigger"
+      - "gcr.io/$PROJECT_ID/release-trigger"
       - "."
 
   - name: gcr.io/gcp-runtimes/structure_test
     args:
-      ["-i", "gcr.io/$PROJECT_ID/kokoro-trigger", "--config", "/workspace/packages/kokoro-trigger/container-test.yaml", "-v"]
+      ["-i", "gcr.io/$PROJECT_ID/release-trigger", "--config", "/workspace/packages/release-trigger/container-test.yaml", "-v"]
     waitFor: ["build-docker"]
