+++ packages/release-trigger/src/config-schema.json
@@ -1,7 +1,7 @@
 {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "title": "kokoro-trigger config",
-    "description": "Schema for defining the kokoro-trigger config",
+    "title": "release-trigger config",
+    "description": "Schema for defining the release-trigger config",
     "additionalProperties": false,
     "type": "object",
     "properties": {
