+++ packages/release-trigger/test/fixtures/config/valid-config.yml
