+++ packages/release-trigger/test/fixtures/events/issue_opened.json
